Add files via upload
IndieStealer is an experimental program made in educational purposes only. It can collect all the data, pack and send it only in memory. This program works with Windows OS. This program does collect all kind of data from your PC. This program is made for pentest and analyze. ALL CREDITS TO https://github.com/jaime-olivares/zipstorer in 
ZipStorer.cs